re-build structure
