--- README.md
@@ -1,1 +1,0 @@
-# cpnt262-a5-JEBX
