+++ server.js
@@ -3,8 +3,12 @@ const mongoose = require('./_connection')
 const dotenv = require('dotenv').config()
 const express = require('express')
 const apiS = require('./routes/api/apiSubscribers')
-// const subscribe = require('./routes/subscribe')
-const index = require('./routes/index')
+// const index = require('./routes/index')
+const sub = require('./routes/subscribe')
+// const team = require('./routes/team')
+// const admin = require('./routes/admin')
+// const gallery = require('./routes/gallery')
+
 
 const app = express()
 
@@ -12,9 +16,21 @@ app.set('view engine', 'ejs');
 
 app.use(express.static('./public'))
 
-app.use('/', index)
+app.use('/subscribe', sub)
+
 app.use('/api/v0', apiS)
-// app.use('/subscribe', subscribe)
+
+app.get('/', (req, res) => {
+  res.render('../views/pages/index', {pageTitle: "Index"})
+})
+
+app.get('/admin', (req, res) => {
+  res.render('../views/pages/admin', {pageTitle: "Subscribers List"})
+})
+
+// app.use('/gallery', gallery)
+// app.use('/team', team)
+
 
 app.use((req, res) => {
   res.sendStatus(404);
