+++ views/pages/subscribed.ejs
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <title>Subscribe</title>
+		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+		<link rel="stylesheet" href="/css/main.css" />
+  </head>
+  <body>
+    <head>
+
+    </head>
+    <main>
+      
+    </main>
+    <footer>
+
+    </footer>
+  </body>
+</html>
